isolate common test code
